backend/python-recommendation-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.3
pymongo==4.5.0
python-dotenv==1.0.0
